rm iomod dir; impl macros; mv iomod rpc to lib
iomod directory is superseded by stdlib-iomod repo
