--- backends/aws-lambda/iomod/dynamodb/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "assemblylift-iomod-dynamodb"
-version = "0.1.0"
-description = "AssemblyLift DynamoDB IOmod (host)"
-authors = ["Akkoro and the AssemblyLift contributors <assemblylift@akkoro.io>"]
-edition = "2018"
-license-file = "../../../../LICENSE.md"
-repository = "https://github.com/akkoro/assemblylift"
-readme = "README.md"
-
-[[bin]]
-name = "dynamodb"
-path = "src/main.rs"
-
-[dependencies]
-capnp = "0.13.3"
-capnp-rpc = "0.13.1"
-
-tokio = { version = "0.2.22", features = ["full"] }
-tokio-util = { version = "0.3.1", features = ["compat"] }
-futures = "0.3.5"
-futures-util = "0.3.5"
-
-rusoto_core = { version = "0.43.0" }
-rusoto_dynamodb = { version = "0.43.0", features = ["serialize_structs", "deserialize_structs"] }
-
-assemblylift_core = { version = "0.2.0", package = "assemblylift-core", path = "../../../../core"}
-assemblylift_core_io = { version = "0.2.0", package = "assemblylift-core-io", path = "../../../../core/io" }
-assemblylift_core_iomod = { version = "0.1.0", package = "assemblylift-core-iomod", path = "../../../../core/iomod" }
-
-wasmer_runtime_core = { version = "0.1.1", package = "wasmer-runtime-core-asml-fork" }
-
-once_cell = "1.4.1"
-serde_json = "1.0.57"
-crossbeam-utils = "0.7.2"
-paste = "0.1.18"
