--- backends/aws-lambda/iomod/dynamodb/guest/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "assemblylift-iomod-dynamodb-guest"
-version = "0.1.0"
-description = "AssemblyLift DynamoDB IOmod (guest)"
-authors = ["Akkoro and the AssemblyLift contributors <assemblylift@akkoro.io>"]
-edition = "2018"
-license-file = "../../../../../LICENSE.md"
-repository = "https://github.com/akkoro/assemblylift"
-readme = "README.md"
-
-[dependencies]
-serde = "1.0.106"
-serde_json = "1.0.53"
-bytes = "0.5.4"
-base64 = "0.12.1"
-
-assemblylift_core_io_guest = { version = "0.1.0", package = "assemblylift-core-io-guest", path = "../../../../../core/io/guest" }
-assemblylift_core_iomod_guest = { version = "0.1.0", package = "assemblylift-core-iomod-guest", path = "../../../../../core/iomod/guest" }
