+++ core/iomod/Cargo.toml
@@ -14,6 +14,7 @@ build = "build.rs"
 tokio = { version = "0.2.21", features = ["full"] }
 tokio-util = { version = "0.3.1", features = ["compat"] }
 futures = "0.3.5"
+futures-util = "0.3.5"
 crossbeam-utils = "0.7.2"
 once_cell = "1.4.0"
 lazy_static = "1.4.0"
